requirements.txt
urllib3==1.22
requests==2.18.4
redis==2.10.6
google==3.0.0
discord.py==1.4.1
python-dotenv==0.14.0
async-timeout==3.0.1